requirements.txt
streamlit
requests
pandas
scikit-learn
joblib
folium
streamlit-folium
plotly
reportlab
python-dotenv